requirements.txt
asgiref==3.8.1
Django==5.1.4
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
phonenumbers==8.13.52
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
